MouseManager Example added
Added example in Scenes/Delete of MouseManagerExample.scene
Added Scripts folder with a folder to contain the Manager scripts.
Can click , hold, drag, and release objects. But, for some reason when I'm in game mode it renders every previous position of the bricks, but if I switch tabs and back, it is fine.